feat(charts): combined into one screen
